feat: rename all models to lowercase
