+++ src/entity/index.ts
@@ -1,8 +1,8 @@
-export * from './Category';
-export * from './User';
-export * from './Comment';
-export * from './NFT';
-export * from './Cart';
-export * from './Artwork';
-export * from './Follow';
-export * from './Feedback';
+export * from './category';
+export * from './user';
+export * from './comment';
+export * from './nft';
+export * from './cart';
+export * from './artwork';
+export * from './follow';
+export * from './feedback';
